requirements.txt
alive-progress
# faiss-cpu
google-api-python-client
jq
# langchain # Automatically installs langchain-community
# langchain-openai
# langsmith
openai
pinecone-client
semantic-kernel
streamlit
tiktoken
youtube-transcript-api